==> code/4_analysis_aim4_ahuang.ipynb <==
### Setup

# install packages
import Pkg;
Pkg.add("CSV");
Pkg.add("DataFrames");
Pkg.add("FreqTables");
Pkg.add("StatsBase");

# load packages
using CSV;
using DataFrames;
using FreqTables;
using StatsBase;

# setup filepaths
path_source = string(@__DIR__,"\\..\\source");
path_dev = string(@__DIR__,"\\..\\dev");
path_output = string(@__DIR__,"\\..\\output");

### Read in dataset

# read in cleaned combined VAERS file
df = CSV.read(joinpath(path_dev,"19-21VAERSCOMB_clean.csv"), DataFrame);
names(df)

size(df)

first(select(df, ["VAERS_ID", "VAX_NAME", "SYMPTOMS", "SERIOUS_EVENT"]), 10)

# Convert SYMPTOMS strings back to arrays of strings
a = fill([], size(df,1));

for rownumber in 1:size(df, 1)
    str_symptoms = df[rownumber,:SYMPTOMS]
    str_index_start = findfirst(isequal('['), str_symptoms)+2
    str_index_end = findfirst(isequal(']'), str_symptoms)-2
    a[rownumber] = split(df[rownumber,:SYMPTOMS][str_index_start:str_index_end], "\", \"")
end;

df.SYMPTOMS = a;
first(select(df, ["VAERS_ID", "VAX_NAME", "SYMPTOMS", "SERIOUS_EVENT"]), 10)

### Exploratory Data Analysis

# Frequency table of vaccines
freq_vax = sort(freqtable(df, :VAX_NAME), rev=true);
freq_vax

# Create dictionary for VAX_NAME:SYMPTOM
# Step 1: create the dict 
vax_to_symptoms_dict = Dict{String, Set{String}}()
# Step 2: populate the keys (VAERS_ID) of the dict
for rownumber in 1:size(df, 1)
    vax_name = df[rownumber, :VAX_NAME]
    if !haskey(vax_to_symptoms_dict, vax_name)
        # this is the set where we will store all of the symptoms for this VAX_NAME
        vax_to_symptoms_dict[vax_name] = Set{String}()
    end
end
# Step 3: populate the values (SYMPTOMS) of the dict
for rownumber in 1:size(df, 1)
    vax_name = df[rownumber, :VAX_NAME]
    symptoms = df[rownumber, :SYMPTOMS]
    for symptom in symptoms 
        push!(vax_to_symptoms_dict[vax_name], symptom)
    end
end
# View dict
vax_to_symptoms_dict

### Among all 112 vaccines, what are the top reported symptoms:

# Create dict of most reported symptoms for all 112 vaccines
symptoms_all_dupes = reduce(vcat, df.SYMPTOMS)
symptoms_freq_dict = StatsBase.countmap(symptoms_all_dupes)
list_symptoms = unique(reduce(vcat, df.SYMPTOMS))
freq_per_symptom = [symptoms_freq_dict[val] for val in list_symptoms]

# print table
tbl_symptoms_freq = hcat(list_symptoms, freq_per_symptom)
tbl_symptoms_freq[sortperm(tbl_symptoms_freq[:, 2]; rev=true), :]

### Among the 3 COVID-19 vaccines, what are the top reported symptoms:

# Create dict of most reported symptoms for COVID vaccines
df_covid19 = filter(row -> row.VAX_TYPE in ["COVID19"], df)

symptoms_all_dupes = reduce(vcat, df_covid19.SYMPTOMS)
symptoms_freq_dict = StatsBase.countmap(symptoms_all_dupes)
list_symptoms = unique(reduce(vcat, df_covid19.SYMPTOMS))
freq_per_symptom = [symptoms_freq_dict[val] for val in list_symptoms]

# print table
tbl_symptoms_freq = hcat(list_symptoms, freq_per_symptom)
tbl_symptoms_freq[sortperm(tbl_symptoms_freq[:, 2]; rev=true), :]

### Define functions for AIM analysis

function get_freqtable(df, vax_name_str, symptom_str)
    test = select(df, ["VAX_NAME", "SYMPTOMS"])
    # Create dummy variable for vax_name
    test.VAX_IND = (test.VAX_NAME .== vax_name_str)
    # Create dummy variable for symptom
    test.SYMPTOM_IND = zeros(size(test, 1))
    for rownumber in 1:size(test, 1)
        if symptom_str in test[rownumber,:SYMPTOMS]
            test[rownumber,:SYMPTOM_IND] = 1
        end
    end
    
    # Create frequency table
    tbl = freqtable(test, :VAX_IND, :SYMPTOM_IND)
    return tbl
end;

function get_PRR(tbl)
    # Calculate PRR
    a = tbl[2,2]
    b = tbl[2,1]
    c = tbl[1,2]
    d = tbl[1,1]
    PRR = (a/(a+b))/(c/(c+d))
    
    return PRR
end;

get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Headache")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Headache"))

get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Headache")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Headache"))

### Contingency tables of Serious Events

# Frequency table of serious events
freq_serious = sort(freqtable(df, :SERIOUS_EVENT), rev=true);
freq_serious

# Example contingency tables for VAX of interest = "COVID19 (COVID19 (PFIZER-BIONTECH))" and serious report
df.COVID19_PFIZER = (df.VAX_NAME .== "COVID19 (COVID19 (PFIZER-BIONTECH))");
tbl = freqtable(df, :COVID19_PFIZER, :SERIOUS_EVENT)

get_PRR(tbl)

df.COVID19_MODERNA = (df.VAX_NAME .== "COVID19 (COVID19 (MODERNA))");
tbl = freqtable(df, :COVID19_MODERNA, :SERIOUS_EVENT)

get_PRR(tbl)

df.COVID19_JANSSEN = (df.VAX_NAME .== "COVID19 (COVID19 (JANSSEN))");
tbl = freqtable(df, :COVID19_JANSSEN, :SERIOUS_EVENT)

get_PRR(tbl)

### Stratified analysis by Age Group

freqtable(df, :AGE_YRS)

# Create the following age bins, following RI eligibility groups
# 16 to 39
df.AGE_16to39 = (df.AGE_YRS .>= 16) .& (df.AGE_YRS .<= 39);

# 40 to 49
df.AGE_40to49 = (df.AGE_YRS .>= 40) .& (df.AGE_YRS .<= 49);

# 50 to 59
df.AGE_50to59 = (df.AGE_YRS .>= 50) .& (df.AGE_YRS .<= 59);

# 60 to 64
df.AGE_60to64 = (df.AGE_YRS .>= 60) .& (df.AGE_YRS .<= 64);

# 65 to 74
df.AGE_65to74 = (df.AGE_YRS .>= 65) .& (df.AGE_YRS .<= 74);

# 75+
df.AGE_75ovr = (df.AGE_YRS .>= 75);

# Create PRRs for SERIOUS_EVENT, stratified by COVID vax, stratified by AGE GROUP
AGE_GROUP = ["16 to 39", "40 to 49", "50 to 59", "60 to 64", "65 to 74", "75+", "All (16 to 75+)"];

# Pfizer
COVID19_PFIZER_PRR = zeros(0);
push!(COVID19_PFIZER_PRR, get_PRR(freqtable(df[df.AGE_16to39,:], :COVID19_PFIZER, :SERIOUS_EVENT)));
push!(COVID19_PFIZER_PRR, get_PRR(freqtable(df[df.AGE_40to49,:], :COVID19_PFIZER, :SERIOUS_EVENT)));
push!(COVID19_PFIZER_PRR, get_PRR(freqtable(df[df.AGE_50to59,:], :COVID19_PFIZER, :SERIOUS_EVENT)));
push!(COVID19_PFIZER_PRR, get_PRR(freqtable(df[df.AGE_60to64,:], :COVID19_PFIZER, :SERIOUS_EVENT)));
push!(COVID19_PFIZER_PRR, get_PRR(freqtable(df[df.AGE_65to74,:], :COVID19_PFIZER, :SERIOUS_EVENT)));
push!(COVID19_PFIZER_PRR, get_PRR(freqtable(df[df.AGE_75ovr,:], :COVID19_PFIZER, :SERIOUS_EVENT)));
push!(COVID19_PFIZER_PRR, get_PRR(freqtable(df, :COVID19_PFIZER, :SERIOUS_EVENT)));

# Moderna
COVID19_MODERNA_PRR = zeros(0);
push!(COVID19_MODERNA_PRR, get_PRR(freqtable(df[df.AGE_16to39,:], :COVID19_MODERNA, :SERIOUS_EVENT)));
push!(COVID19_MODERNA_PRR, get_PRR(freqtable(df[df.AGE_40to49,:], :COVID19_MODERNA, :SERIOUS_EVENT)));
push!(COVID19_MODERNA_PRR, get_PRR(freqtable(df[df.AGE_50to59,:], :COVID19_MODERNA, :SERIOUS_EVENT)));
push!(COVID19_MODERNA_PRR, get_PRR(freqtable(df[df.AGE_60to64,:], :COVID19_MODERNA, :SERIOUS_EVENT)));
push!(COVID19_MODERNA_PRR, get_PRR(freqtable(df[df.AGE_65to74,:], :COVID19_MODERNA, :SERIOUS_EVENT)));
push!(COVID19_MODERNA_PRR, get_PRR(freqtable(df[df.AGE_75ovr,:], :COVID19_MODERNA, :SERIOUS_EVENT)));
push!(COVID19_MODERNA_PRR, get_PRR(freqtable(df, :COVID19_MODERNA, :SERIOUS_EVENT)));

# Janssen
COVID19_JANSSEN_PRR = zeros(0);
push!(COVID19_JANSSEN_PRR, get_PRR(freqtable(df[df.AGE_16to39,:], :COVID19_JANSSEN, :SERIOUS_EVENT)));
push!(COVID19_JANSSEN_PRR, get_PRR(freqtable(df[df.AGE_40to49,:], :COVID19_JANSSEN, :SERIOUS_EVENT)));
push!(COVID19_JANSSEN_PRR, get_PRR(freqtable(df[df.AGE_50to59,:], :COVID19_JANSSEN, :SERIOUS_EVENT)));
push!(COVID19_JANSSEN_PRR, get_PRR(freqtable(df[df.AGE_60to64,:], :COVID19_JANSSEN, :SERIOUS_EVENT)));
push!(COVID19_JANSSEN_PRR, get_PRR(freqtable(df[df.AGE_65to74,:], :COVID19_JANSSEN, :SERIOUS_EVENT)));
push!(COVID19_JANSSEN_PRR, get_PRR(freqtable(df[df.AGE_75ovr,:], :COVID19_JANSSEN, :SERIOUS_EVENT)));
push!(COVID19_JANSSEN_PRR, get_PRR(freqtable(df, :COVID19_JANSSEN, :SERIOUS_EVENT)));

tbl_aim4 = hcat(AGE_GROUP, COVID19_PFIZER_PRR, COVID19_MODERNA_PRR, COVID19_JANSSEN_PRR)

df_aim4 = DataFrame(tbl_aim4, Symbol.(["AGE GROUP", "PFIZER-BIONTECH", "MODERNA", "JANSSEN"]))

# write to output folder
CSV.write(joinpath(path_output,"aim4.csv"), df_aim4);

### Janssen and Blood Clots?

#### Thrombocytopenia

get_freqtable(df, "COVID19 (COVID19 (JANSSEN))", "Thrombocytopenia")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (JANSSEN))", "Thrombocytopenia"))

get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Thrombocytopenia")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Thrombocytopenia"))

get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Thrombocytopenia")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Thrombocytopenia"))

#### Deep vein thrombosis

get_freqtable(df, "COVID19 (COVID19 (JANSSEN))", "Deep vein thrombosis")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (JANSSEN))", "Deep vein thrombosis"))

get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Deep vein thrombosis")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Deep vein thrombosis"))

get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Deep vein thrombosis")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Deep vein thrombosis"))

#### Cerebral venous sinus thrombosis

get_freqtable(df, "COVID19 (COVID19 (JANSSEN))", "Cerebral venous sinus thrombosis")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (JANSSEN))", "Cerebral venous sinus thrombosis"))

get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Cerebral venous sinus thrombosis")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (PFIZER-BIONTECH))", "Cerebral venous sinus thrombosis"))

get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Cerebral venous sinus thrombosis")

get_PRR(get_freqtable(df, "COVID19 (COVID19 (MODERNA))", "Cerebral venous sinus thrombosis"))

#### Women Age 16-39

get_freqtable(df[df.AGE_16to39,:], "COVID19 (COVID19 (JANSSEN))", "Cerebral venous sinus thrombosis")

get_PRR(get_freqtable(df[df.AGE_16to39,:], "COVID19 (COVID19 (JANSSEN))", "Cerebral venous sinus thrombosis"))

get_freqtable(df[df.AGE_16to39,:], "COVID19 (COVID19 (PFIZER-BIONTECH))", "Cerebral venous sinus thrombosis")

get_PRR(get_freqtable(df[df.AGE_16to39,:], "COVID19 (COVID19 (PFIZER-BIONTECH))", "Cerebral venous sinus thrombosis"))

get_freqtable(df[df.AGE_16to39,:], "COVID19 (COVID19 (MODERNA))", "Cerebral venous sinus thrombosis")

get_PRR(get_freqtable(df[df.AGE_16to39,:], "COVID19 (COVID19 (MODERNA))", "Cerebral venous sinus thrombosis"))
